+++ biopytools/subseq/__init__.py
@@ -1,0 +1,8 @@
+"""
+序列子集提取模块 | Sequence Subsequence Extraction Module
+"""
+
+from .core import SequenceExtractor
+from .utils import SubseqLogger
+
+__all__ = ['SequenceExtractor', 'SubseqLogger']